intermediate revision - massive refactoring
