+++ README.md
@@ -1,38 +1,82 @@
-# JNAutoIt - Java Native Access for AutoIt
+### Info
 
+This directory contains a replica of [JNAutoIt - JNA AutoitX dll wrapper](https://github.com/midorlo/JNAutoIt)
+by Midorlo that generates a java method wrapping all core AutoitX [functions](https://documentation.help/AutoItX/)
+from the dll which can be e.g. installed from
+[Nuget package for AutoitX](https://www.nuget.org/packages/AutoItX/).
+without doing the COM registration. This project is also tracked in a [fork](https://github.com/sergueik/JNAutoIt)
+of the genuine [JNAutoIt - Java Native Access for AutoIt](https://github.com/midorlo/JNAutoIt)
+but since code differences are somewhat big, no upstream PR is available at this time.
 
-![Build](https://img.shields.io/appveyor/ci/midorlo/JNAutoit.svg) ![License](https://img.shields.io/badge/license-APACHE-blue.svg)
+### Project Goals
 
-
-## Basic Overview
-
-
-This project aims to bring AutoIt functionality to Java. To do so, it wraps [AutoItX](https://documentation.help/AutoItX/) via [Java Native Access](https://github.com/java-native-access/jna). It should serve as a successor of [accessrichard/autoitx4java](https://github.com/accessrichard/autoitx4java). 
-
-
-## Building from source
-
-```maven
-mvn clean install
+The project adds a more java-friendly methods like
+```java
+public String WinGetText(String title, String text) {
+// ...
+}
 ```
 
-This will require a windows >=8 host. 
+compared to underlying
+```java
+public void AU3_WinGetText(WString title, WString text, LPWSTR resultPointer, int bufSize) {
+// ...	
+}
+```
+both wrapping the [WinGetText](https://www.autoitscript.com/autoit3/docs/functions/WinGetText.htm) method for retrieving the text from a window.
+```basic
+WinGetText
+Retrieves the text from a window.
 
-### Dependencies
+WinGetText ( "title" [, "text"] )
+```
+Currently converted to java-friendly signatures:
 
-* [Java Native Access(bundled via Maven)](https://github.com/java-native-access/jna) 
-* [TestNG(bundled via Maven)](https://testng.org/doc/index.html)
-* [AutoItX(Bundled)](https://www.autoitscript.com/site/autoit/downloads/)
+* `WinClose`
+* `WinGetText`
+* `WinWaitActive`
+* `Send`
 
+This addresses the anticipated needs of AutoIt with Selenium testing.
 
-## Usage
+Converting all methods is a work in progress (adequate tests might be a  bit of a challenge).
 
+Sending Control + regular key combinations to the application e.g. to control browser zoom is is a work in progress.
+### Example Usage
+
+Provide Path to save the downloaded File. This is much lightweight compared to the
+original  [file Upload Example using AutoIT and with Selenium Webdriver](https://www.guru99.com/use-autoit-selenium.html) recipe.
+which requires an full "compile script to exe"
+Windows executable generation from the following sample `.au3` [script](https://automated-testing.info/t/webdriver-features-rabota-s-upload-popup-windows-native-okno-pri-pomoshhi-selenium-web-driver/2288):
+```basic
+WinWaitActive("Open") // ждем активности Windows окна Open
+Send("D:\AutoIT-commands\TestingVideo.mp4") //Отправляем иму путь к файлу (фокус по умолчанию стоит на текст-боксе где прописываеться путь к файлу)
+Send("{ENTER}") //эмулируем нажатие клавиши Enter (а-ля загрузить)
+```
 
 ```java
-import de.midorlo.jnautoit.jna.AutoItX;
-..
-AutoItX aix = AutoItX.getInstance();
-aix.AU3_MouseMove(640, 480);
+import example.AutoItX;
+// pending PR
+// import de.midorlo.jnautoit.jna.AutoItX;
+
+  String windowTitle = "Save";
+  String windowText = "";
+  AutoItX instance = AutoItX.getInstance();
+  String filePath =	"C:\\Users\\user\\Downloads\\TestingVideo.mp4";
+  // Selenium test launches file download
+  instance.WinWaitActive(windowTitle, windowText);
+  instance.Send(filePath);
+  instance.Send("\n");
 ```
 
-        
+### See Also
+  * [Index of /autoit3/docs/functions](https://www.autoitscript.com/autoit3/docs/functions/)
+  * Powershell iAutoIt cmdlets [documentation](https://www.autoitconsulting.com/site/scripting/autoit-cmdlets-for-windows-powershell/)
+  * [Autoit forum (in Russian)](http://autoit-script.ru/index.php).
+  * [Misc. autoit-related topics in QA forum (in Russian, mostly)](https://automated-testing.info/search?q=autoit)
+
+### License
+This project is licensed under the terms of the MIT license.
+
+### Author
+[Serguei Kouzmine](kouzmine_serguei@yahoo.com)
